Got cache reads working for circ supply and refactored common cache
