+++ nym-api/src/circulating_supply_api/cache/data.rs
@@ -1,0 +1,13 @@
+use crate::caching_support::Cache;
+
+pub(crate) struct CirculatingSupplyCacheData {
+    pub(crate) circulating_supply: Cache<String>,
+}
+
+impl CirculatingSupplyCacheData {
+    pub fn new() -> CirculatingSupplyCacheData {
+        CirculatingSupplyCacheData {
+            circulating_supply: Cache::default(),
+        }
+    }
+}
