+++ nym-api/src/circulating_supply_api/cache/mod.rs
@@ -1,0 +1,45 @@
+use rocket::fairing::AdHoc;
+use std::{
+    sync::{atomic::AtomicBool, Arc},
+    time::Duration,
+};
+use tokio::sync::RwLock;
+use tokio::time;
+
+use crate::caching_support::Cache;
+
+use self::data::CirculatingSupplyCacheData;
+
+mod data;
+mod refresher;
+
+#[derive(Clone)]
+pub(crate) struct CirculatingSupplyCache {
+    initialised: Arc<AtomicBool>,
+    data: Arc<RwLock<CirculatingSupplyCacheData>>,
+}
+
+impl CirculatingSupplyCache {
+    fn new() -> CirculatingSupplyCache {
+        CirculatingSupplyCache {
+            initialised: Arc::new(AtomicBool::new(false)),
+            data: Arc::new(RwLock::new(CirculatingSupplyCacheData::new())),
+        }
+    }
+
+    pub(crate) async fn say_foomp(&self) -> Option<Cache<String>> {
+        match time::timeout(Duration::from_millis(100), self.data.read()).await {
+            Ok(cache) => Some(cache.circulating_supply.clone()),
+            Err(e) => {
+                error!("{}", e);
+                Some(Cache::new(String::from("0nym")))
+            }
+        }
+    }
+
+    pub(crate) fn stage() -> AdHoc {
+        AdHoc::on_ignite("Circulating Supply Cache Stage", |rocket| async {
+            rocket.manage(Self::new())
+        })
+    }
+}
