+++ nym-api/src/node_status_api/cache/mod.rs
@@ -1,9 +1,10 @@
 // Copyright 2022 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
+use crate::caching_support::Cache;
+
 use self::data::NodeStatusCacheData;
 use self::inclusion_probabilities::InclusionProbabilities;
-use crate::contract_cache::cache::Cache;
 use mixnet_contract_common::MixId;
 use nym_api_requests::models::{MixNodeBondAnnotated, MixnodeStatus};
 use rocket::fairing::AdHoc;
